Unit tests for basic floating-point tensor functionality completed.
